refactor: extract components to new file
